Archetectural update Divide logic into a separate module Devide parser and apt interaction modules
